fix: Resoleve build warning
size_t -> int32 변환으로 인해 발생하는 경고 막기 위해 uint64로 변경

+++ TL2/FSceneRenderer.cpp
@@ -317,8 +317,8 @@ void FSceneRenderer::FinalizeFrame()
 	if (World->GetRenderSettings().IsShowFlagEnabled(EEngineShowFlags::SF_Culling))
 	{
 		int totalActors = static_cast<int>(World->GetActors().size());
-		int visiblePrimitives = Proxies.Primitives.size();
-		UE_LOG("Total Actors: %d, Visible Primitives: %d\r\n", totalActors, visiblePrimitives);
+		uint64 visiblePrimitives = Proxies.Primitives.size();
+		UE_LOG("Total Actors: %d, Visible Primitives: %llu\r\n", totalActors, visiblePrimitives);
 	}
 }
